feat: creating implementation of repository

+++ src/main/java/com/streaming/app/streaming/catalog_item/domain/CatalogItemRepository.java
@@ -6,7 +6,7 @@ public interface CatalogItemRepository {
 
     public void save(CatalogItem catalogItem);
 
-    public int count();
+    public Long count();
 
     public CatalogItem findById(CatalogItemId id);
